+++ src/tasks/buildTaskProvider.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
+import * as semver from 'semver';
 import { CommandTaskProvider, TitaniumTaskBase, TitaniumTaskDefinitionBase, TitaniumBuildBase } from './commandTaskProvider';
 import { selectBuildTarget } from '../quickpicks/build/common';
 import { TaskExecutionContext, debugSessionInformation, DEBUG_SESSION_VALUE } from './tasksHelper';
@@ -9,6 +10,8 @@ import { TaskPseudoTerminal } from './taskPseudoTerminal';
 import { Platform } from '../types/common';
 import { Command } from './commandBuilder';
 import { getValidWorkspaceFolders, promptForWorkspaceFolder } from '../quickpicks';
+import { ExtensionContainer } from '../container';
+import { DevelopmentTarget } from '../types/cli';
 
 export interface BuildTask extends TitaniumTaskBase {
 	definition: BuildTaskDefinitionBase;
@@ -22,6 +25,14 @@ export interface AppBuildTaskDefinitionBase extends TitaniumTaskDefinitionBase {
 	titaniumBuild: AppBuildTaskTitaniumBuildBase;
 }
 
+export interface ModuleBuildTask extends BuildTask {
+	definition: ModuleBuildTaskDefinitionBase;
+}
+
+export interface ModuleBuildTaskDefinitionBase extends TitaniumTaskDefinitionBase {
+	titaniumBuild: ModuleBuildTaskTitaniumBuildBase;
+}
+
 export interface BuildTaskDefinitionBase extends TitaniumTaskDefinitionBase {
 	titaniumBuild: AppBuildTaskTitaniumBuildBase | ModuleBuildTaskTitaniumBuildBase;
 }
@@ -33,7 +44,7 @@ export interface BuildTaskTitaniumBuildBase extends TitaniumBuildBase {
 
 export interface AppBuildTaskTitaniumBuildBase extends BuildTaskTitaniumBuildBase {
 	deviceId?: string;
-	target?: 'device' | 'emulator' | 'simulator';
+	target?: DevelopmentTarget;
 	projectType?: 'app';
 	liveview?: boolean;
 	deployType?: 'development' | 'test';
@@ -42,7 +53,19 @@ export interface AppBuildTaskTitaniumBuildBase extends BuildTaskTitaniumBuildBas
 }
 
 export interface ModuleBuildTaskTitaniumBuildBase extends BuildTaskTitaniumBuildBase {
+	deviceId?: string;
+	target?: DevelopmentTarget;
 	projectType?: 'module';
+}
+
+function sdkSupportsModuleBuildTargets (): boolean {
+	const sdk = ExtensionContainer.environment.selectedSdk();
+
+	if (sdk && semver.gte(sdk.version, '12.1.0')) {
+		return true;
+	}
+
+	return false;
 }
 
 export class BuildTaskProvider extends CommandTaskProvider {
@@ -126,7 +149,12 @@ export class BuildTaskProvider extends CommandTaskProvider {
 
 			return helper.resolveAppBuildCommandLine(context, task.definition.titaniumBuild);
 		} else if (definition.titaniumBuild.projectType === 'module') {
+
 			definition.titaniumBuild.projectDir = path.join(definition.titaniumBuild.projectDir, definition.titaniumBuild.platform);
+
+			if (!definition.titaniumBuild.target && sdkSupportsModuleBuildTargets()) {
+				definition.titaniumBuild.target = (await selectBuildTarget(definition.titaniumBuild.platform)).id as 'device' | 'emulator' | 'simulator';
+			}
 			return helper.resolveModuleBuildCommandLine(context, task.definition.titaniumBuild);
 		} else {
 			throw new Error(`Unknown project type ${definition.titaniumBuild.projectType}`);