+++ src/types/cli.ts
@@ -18,9 +18,9 @@ export interface CreateModuleOptions extends CreateOptions {
 }
 
 export type Target = DevelopmentTarget | DeploymentTarget;
-export type DevelopmentTarget = 'device' | 'emulator' | 'simulator';
-export type DeploymentTarget = 'dist-adhoc' | 'dist-appstore' | 'dist-playstore';
+export type DevelopmentTarget = 'device' | 'emulator' | 'macos' | 'simulator';
+export type DeploymentTarget = 'dist-adhoc' | 'dist-appstore' | 'dist-macappstore' | 'dist-playstore';
 
 export type PrettyTarget = PrettyDevelopmentTarget | PrettyDeploymentTarget;
-export type PrettyDevelopmentTarget = 'Device' | 'Emulator' | 'Simulator';
-export type PrettyDeploymentTarget = 'Ad-Hoc' | 'App Store' | 'Play Store';
+export type PrettyDevelopmentTarget = 'Device' | 'Emulator' | 'Simulator' | 'macOS';
+export type PrettyDeploymentTarget = 'Ad-Hoc' | 'App Store' | 'macOS App Store' | 'Play Store';